// src/types/informationEntrieShopTypes.ts
export interface InformationEntrieShop {
  distance: number;
  sourceAddress: string;
  destinationAddress: string;
  sourceLongitude: number;
  sourceLatitude: number;
  destinationLongitude: number;
  destinationLatitude: number;
  destinationProvince: string;
  destinationTel: string;
  destinationContact: string;
}